Conflicts:
	system/src/de/uka/ilkd/key/logic/TermBuilder.java
	system/src/de/uka/ilkd/key/proof/init/AbstractOperationPO.java
	system/src/de/uka/ilkd/key/proof/init/FunctionalOperationContractPO.java
	system/src/de/uka/ilkd/key/proof/mgt/SpecificationRepository.java
	system/src/de/uka/ilkd/key/rule/BlockContractRule.java
	system/src/de/uka/ilkd/key/rule/DefaultBuiltInRuleApp.java
	system/src/de/uka/ilkd/key/rule/LoopInvariantBuiltInRuleApp.java
	system/src/de/uka/ilkd/key/rule/QueryExpand.java
	system/src/de/uka/ilkd/key/rule/Taclet.java
	system/src/de/uka/ilkd/key/rule/UseOperationContractRule.java
	system/src/de/uka/ilkd/key/rule/WhileInvariantRule.java
	system/src/de/uka/ilkd/key/rule/metaconstruct/ExpandQueriesMetaConstruct.java
	system/src/de/uka/ilkd/key/speclang/ContractFactory.java
	system/src/de/uka/ilkd/key/speclang/LoopInvariantImpl.java

+++ system/src/de/uka/ilkd/key/proof/Proof.java
@@ -1,20 +1,27 @@
-// This file is part of KeY - Integrated Deductive Software Design
+// This file is part of KeY - Integrated Deductive Software Design 
 //
-// Copyright (C) 2001-2011 Universitaet Karlsruhe (TH), Germany
+// Copyright (C) 2001-2011 Universitaet Karlsruhe (TH), Germany 
 //                         Universitaet Koblenz-Landau, Germany
 //                         Chalmers University of Technology, Sweden
-// Copyright (C) 2011-2013 Karlsruhe Institute of Technology, Germany
+// Copyright (C) 2011-2013 Karlsruhe Institute of Technology, Germany 
 //                         Technical University Darmstadt, Germany
 //                         Chalmers University of Technology, Sweden
 //
-// The KeY system is protected by the GNU General
+// The KeY system is protected by the GNU General 
 // Public License. See LICENSE.TXT for details.
-//
-
+// 
 
 package de.uka.ilkd.key.proof;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.Iterator;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.Vector;
 
 import de.uka.ilkd.key.collection.ImmutableList;
 import de.uka.ilkd.key.collection.ImmutableSLList;
@@ -24,7 +31,13 @@ import de.uka.ilkd.key.gui.configuration.ProofSettings;
 import de.uka.ilkd.key.gui.configuration.SettingsListener;
 import de.uka.ilkd.key.java.JavaInfo;
 import de.uka.ilkd.key.java.Services;
-import de.uka.ilkd.key.logic.*;
+import de.uka.ilkd.key.logic.Name;
+import de.uka.ilkd.key.logic.Named;
+import de.uka.ilkd.key.logic.NamespaceSet;
+import de.uka.ilkd.key.logic.Semisequent;
+import de.uka.ilkd.key.logic.Sequent;
+import de.uka.ilkd.key.logic.SequentFormula;
+import de.uka.ilkd.key.logic.Term;
 import de.uka.ilkd.key.pp.AbbrevMap;
 import de.uka.ilkd.key.proof.Node.NodeIterator;
 import de.uka.ilkd.key.proof.init.InitConfig;
@@ -254,8 +267,12 @@ public class Proof implements Named {
         }
 
         // Do required cleanup
-        services.getSpecificationRepository().removeProof(this);
-        localMgt.removeProofListener(); // This is strongly required because the listener is contained in a static List
+        if (services != null) {
+           services.getSpecificationRepository().removeProof(this);
+        }
+        if (localMgt != null) {
+           localMgt.removeProofListener(); // This is strongly required because the listener is contained in a static List
+        }
         // remove setting listener from settings
         setSettings(null);
         // set every reference (except the name) to null
@@ -1007,7 +1024,7 @@ public class Proof implements Named {
      * This implementation traverses the proof tree only once.
      */
     public List<Pair<String,String>> statistics() {
-        final List<Pair<String,String>> res = new ArrayList<Pair<String,String>>();
+        final List<Pair<String,String>> res = new ArrayList<Pair<String,String>>(10);
         final NodeIterator it = root().subtreeIterator();
 
         int nodes = 0;
